+++ www/install/install.sql
@@ -187,7 +187,7 @@ CREATE TABLE `core_user` (
   `start_module` varchar(50) NOT NULL DEFAULT 'summary',
   `language` varchar(20) NOT NULL DEFAULT 'en',
   `theme` varchar(20) NOT NULL DEFAULT 'Default',
-  `firstWeekday` tinyint(4) NOT NULL DEFAULT 0,
+  `firstWeekday` tinyint(4) NOT NULL DEFAULT 1,
   `sort_name` varchar(20) NOT NULL DEFAULT 'first_name',
   `muser_id` int(11) NOT NULL DEFAULT 0,
   `mute_sound` tinyint(1) NOT NULL DEFAULT 0,