+++ www/views/Extjs3/javascript/go/Window.js
@@ -66,11 +66,5 @@ go.Window = Ext.extend(Ext.Window, {
 			if(pos[0]<0 || pos[0]+this.width>vpW || pos[1]<0 || pos[1]+this.height>vpH)
 				this.center();
 		}
-	},
-	
-	render : function(container, position){
-		// Fix for chrome inspector
-		container = Ext.get("window-container");
-		return go.Window.superclass.render.call(this, container, position);
 	}
 });